Enable movement based parallax only when sustained movement is detected.
Firefox 6 added support for the built-in accelerometer in MacBooks. This
caused the movement based parallaxing overlap with the moused based
parallaxing. This commit tests the movement to see if it's strong enough
to be considered actual movement. Once true movement is detected, Plax
disables mouse based parallax and switches over to movement based.