Detect element position, then override positioning method
This is to fix #15 in a better way, and fix another bug in FF that the
previous fix caused.

Firefox appears to set explicit values for top, right, bottom, and left
values, regardless of what is actually specified. This negates the
ability to use `.css('left') != 'auto'`. This method, rather than
detecting how the element position is set, figures out where the
element is, then repositions it using top/left based on the original
position.