Add gyroscopic support
This removes a lot of spaghetti code surrounding how we handled
accelerometer support in Plax in favor of using the iPhone's
gyroscope instead. This means we get a clean range-of-motion
effect, much as if you were looking through a window at the object.
Kind of. Sort of.

This probably removes support for desktop (which I think is a good
thing, since people tend not to move their laptops around), and
potentially support for Android and weird devices like that. We can
add that back in later.

In general, though, I think this greatly increases the readability
of the project and gives it an effect that makes a lot more sense.